// frontend/src/pages/Home/interfaces/ServiceDataProps.ts
export interface ServiceDataProps {
  title: string;
  description: string;
  linkText: string;
  image: string;
  link: string;
}

export interface InfoDataProps {
  icon: string,
  title: string;
  description: string;
}

export interface HowItWorksDataProps {
  icon: string;
  title: string;
  description: string;
}